pyproject.toml
[tool.poetry]
name = "ipfs-video-index"
version = "0.1.0"
description = ""
authors = ["Bram Neijt <bram@neijt.nl>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.0"
py-multihash = "^2.0.1"
dqp = "^0.0.3"
aiofiles = "^0.7.0"
plyvel = "^1.3.0"
requests = "^2.26.0"
python-magic = "^0.4.24"
pymetalink = "^6.2"
loguru = "^0.5.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
uvicorn = "^0.14.0"
poetry-lock-package = "^0.4.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
